@@ -1,5 +1,8 @@
 package com.asgatech.sharjahmuseums.Models;
 
+import com.google.gson.annotations.Expose;
+import com.google.gson.annotations.SerializedName;
+
 /**
  * Created by halima.reda on 9/11/2017.
  */
@@ -10,13 +13,26 @@ public class AllSliderModel {
      * Image : Upload/20170822145359436961.png
      */
 
-    private String Image;
+    @SerializedName("ID")
+    @Expose
+    private Integer iD;
+    @SerializedName("Image")
+    @Expose
+    private String image;
 
-    public void setImage(String Image) {
-        this.Image = Image;
+    public Integer getID() {
+        return iD;
+    }
+
+    public void setID(Integer iD) {
+        this.iD = iD;
     }
 
     public String getImage() {
-        return Image;
+        return image;
+    }
+
+    public void setImage(String image) {
+        this.image = image;
     }
 }
